service-workers/service-worker/fetch-event-respond-with-readable-stream.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>respondWith with a response built from a ReadableStream</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
  "use strict";

  const WORKER = "resources/fetch-event-respond-with-readable-stream-worker.js";
  const SCOPE = "resources/blank.html";

  // Register a service worker, then create an iframe at url.
  function iframeTest(url, callback, name) {
    return promise_test(async t => {
      const registration = await service_worker_unregister_and_register(
        t,
        WORKER,
        SCOPE
      );
      await wait_for_state(t, registration.installing, "activated");
      const iframe = await with_iframe(url);
      const contentWindow = iframe.contentWindow;
      t.add_cleanup(async () => {
        if (iframe) iframe.remove();
        if (registration) await registration.unregister();
      });
      await callback(t, contentWindow);
    }, name);
  }

  iframeTest(
    SCOPE,
    async (t, iwin) => {
      const response = await iwin.fetch("?stream");
      assert_equals(await response.text(), "PASS");
    },
    "Subresource built from a ReadableStream"
  );

  iframeTest(
    SCOPE + "?stream",
    (t, iwin) => {
      assert_equals(iwin.document.body.textContent, "PASS");
    },
    "Main resource built from a ReadableStream"
  );

  iframeTest(
    SCOPE,
    async (t, iwin) => {
      const response = await iwin.fetch("?stream&delay");
      assert_equals(await response.text(), "PASS");
    },
    "Subresource built from a ReadableStream - delayed"
  );

  iframeTest(
    SCOPE + "?stream&delay",
    (t, iwin) => {
      assert_equals(iwin.document.body.textContent, "PASS");
    },
    "Main resource built from a ReadableStream - delayed"
  );

  iframeTest(
    SCOPE,
    async (t, iwin) => {
      const response = await iwin.fetch("?stream&use-fetch-stream");
      assert_equals(await response.text(), "PASS\n");
    },
    "Subresource built from a ReadableStream - fetch stream"
  );

  iframeTest(
    SCOPE + "?stream&use-fetch-stream",
    (t, iwin) => {
      assert_equals(iwin.document.body.textContent, "PASS\n");
    },
    "Main resource built from a ReadableStream - fetch stream"
  );
</script>
